requirements.txt
django==4.1.4
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-extensions==3.2.3
drf-spectacular==0.26.4
celery==5.3.1
redis==4.6.0
requests==2.31.0
django-celery-results==2.5.1
